Se añadio input de emojis: - Para esto se creo un nuevo componente "Input emoji" y el mismo  se agrego en todos los lugares que eran necesario. - Tambien se modifico el servicio document-click-service para que ayude al input emoji a mostrarse y interactuar con los diferentes "text input"

+++ src/app/services/document-click.service.ts
@@ -1,19 +1,55 @@
-import { Injectable } from '@angular/core';
+import { ComponentRef, Injectable } from '@angular/core';
+import { InputEmojisComponent } from '../components/input-emojis/input-emojis.component';
 
 @Injectable({
   providedIn: 'root'
 })
 export class DocumentClickService {
 
+  private insertEmojiInput!: HTMLInputElement | HTMLTextAreaElement | undefined;
+  private emojiForm!: InputEmojisComponent;
+
   constructor() { }
 
   formFocusOut(idForm: string, destroyFuncion: () => void){
     const handleFormGroupEvent = (event: Event) => {
       const target = event.target as any;
+      if(target.form?.id == "formEmoji" || target.id == "formEmoji" || target.role == "button"){ 
+        console.log("Emoji form pressed");
+        return;
+      }
       if(target.form?.id != idForm && target.id != idForm && target.tagName != 'HTML'){
         destroyFuncion();
       }
     }
     return handleFormGroupEvent;
+  }
+
+  openEmojiForm(insertIn:HTMLInputElement | HTMLTextAreaElement){
+    const emojiForm: HTMLElement = document.getElementById("formEmoji") as HTMLElement;
+    if(emojiForm != null){
+      emojiForm.style.display = "block";
+      emojiForm.style.top = insertIn.getBoundingClientRect().bottom + "px";
+      emojiForm.style.left = (insertIn.getBoundingClientRect().right - emojiForm.clientWidth) + "px";
+    }
+    this.insertEmojiInput = insertIn;
+    insertIn.focus();
+    this.emojiForm.clickEventConfiguration();
+  }
+
+  closeEmojiForm(){
+    this.insertEmojiInput = undefined;
+    this.emojiForm.closeEmojiPanel();
+  }
+
+  emojiFormClicked(emoji :string){
+    if(this.insertEmojiInput){
+      this.insertEmojiInput.value = this.insertEmojiInput.value + emoji;
+      this.insertEmojiInput.focus();
+    }
+  }
+
+  setEmojiForm(element: InputEmojisComponent){
+    this.emojiForm = element;
   }
 }
